requirements.txt
appdirs==1.4.3
asgiref==3.4.1
binarytree==6.3.0
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Django==3.2.5
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
graphviz==0.16
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
Markdown==3.3.4
msgpack==0.6.2
numpy==1.20.2
packaging==20.3
pep517==0.8.2
progress==1.5
pycodestyle==2.7.0
PyJWT==2.1.0
pyparsing==2.4.6
pytoml==0.1.21
pytz==2021.1
requests==2.22.0
retrying==1.3.3
setuptools-scm==6.0.1
six==1.14.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.25.8
webencodings==0.5.1
